simplified test function
